backend/pyproject.toml
[project]
name = "backend"
version = "0.0.1"
description = "FastAPI backend for Task Management System"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "bcrypt>=4.3.0,<5.0.0",
    "fastapi>=0.121.2",
    "granian>=2.6.0",
    "greenlet>=3.2.4",
    "pydantic[email]>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
test = [
    "aiosqlite>=0.21.0",
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-xdist>=3.8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
]
# PERFORMANCE: Using "auto" mode for convenience
# "auto" has slight overhead but automatically detects async tests
# For better performance, use "strict" mode and add @pytest.mark.asyncio to all tests
# Current tradeoff: convenience over ~5-10% test speed improvement
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
